feat(core, examples): GPU CUDA determinism and auto device selection
Implement GPU-ready broadphase determinism and stress test GPU support:

- **Stable Morton sort (64-bit keys)**: Change d_morton_ to Buffer<uint64_t>
  encoding (morton30 << 32 | body_idx). This makes the bitonic sort stable
  — equal Morton codes are always ordered by body index, giving a fully
  deterministic Karras tree structure across runs.

- **Sequential AABB refit**: Replace parallel_for(n, ...) with
  parallel_for(1, ...) containing a sequential inner loop. Empirical testing
  showed that acquire-release semantics of seq_cst atomics with
  memory_scope::device are insufficient for non-atomic AABB reads/writes
  under CUDA's memory model, causing non-deterministic results. Sequential
  processing on a single work-item guarantees consistent results across
  CUDA/HIP/OpenMP backends.

- **Auto GPU/CPU device selection**: stress_test_blocks now attempts
  Device::default_gpu() and falls back to CPU on failure. Added --cpu flag
  to force CPU-only execution. Device name and type logged at startup.

- **Karras helper updates**: lbvh_delta, find_range, find_split now take
  const uint64_t* and remove the equal-code special case (64-bit keys are
  unique). This simplifies tree construction logic.

- **README updates**: Document stable sort design, sequential refit rationale,
  add RTX 3060 benchmark row, and provide GPU quick-start command.

Resolves foundational GPU determinism gap blocking Phase 1 completion.

+++ core/include/core/broadphase.hpp
@@ -58,7 +58,9 @@ private:
     uint32_t max_pairs_  = 0;
 
     // Bitonic sort input: one entry per body (padded to max_padded_).
-    Buffer<uint32_t> d_morton_;
+    // Key = (morton30 << 32) | body_idx so equal Morton codes sort by body index —
+    // makes the sort stable and the Karras tree deterministic across runs.
+    Buffer<uint64_t> d_morton_;
     Buffer<uint32_t> d_sorted_idx_;  // sorted_idx[k] = original body index of leaf k
 
     // BVH node arrays — total 2*max_bodies-1 nodes:
